// src/shared/models/technologies.ts
export enum Technologies {
  HTML = "HTML",
  CSS = "CSS",
  JavaScript = "JavaScript",
  TypeScript = "TypeScript",
  CSharp = "C#",
  Python = "Python",
  Ruby = "Ruby",
  JQuery = "JQuery",
  React = "React.js",
  Angular = "Angular",
  Vue = "Vue.js",
  Redux = "Redux",
  Node = "Node.js",
  DotNet = "ASP.NET",
  Django = "Django",
  FastAPI = "FastAPI",
  Flask = "Flask",
  ROR = "Ruby on Rails",
  MongoDB = "MongoDB",
  MySQL = "MySQL",
  PostgreSQL = "PostgreSQL",
  Figma = "Figma",
  Jest = "Jest",
  Selenium = "Selenium",
  Pytest = "Pytest",
  AWS = "AWS",
  Docker = "Docker",
  Apache = "Apache",
  Azure = "Azure",
  GraphQL = "GraphQL",
  Heroku = "Heroku",
  Jira = "Jira",
  Git = "Git",
}